Foundations for Space note, properly positioned UI spacing
NOTE: For space note selector, the width, collider width, and object pos are generated in RhythmController when the lane count is changed or loaded. The math is ran based on floats 'backgroundWith' and 'dividerWidth', hard coded referring to the width of the transparent grey background behind every lane (135) and the width of grey dividers separating lanes (15). The space note selector is also not implemented in RhythmRunner or the RhythmDemo scene yet