More Building, but we need to talk...
Made some more organizational gameobjects, put a bunch of different models in. Made a sidewalk with some props like a vending machine and street sign on it near one building to show general feel. Put some plants and buildign sin park, and added more buildings in general. I also put in buildings that were for the previous game idea. While we do lose the whole "It's exactly like the track", the buildings are cool, and we kinda already abandoned accuracy. If you don't want to do it that's chill, but the world looks cooler with them in.

The problem is with scaling.  Currently our player is way too big for the scaling to make sense, so either we scale the plaeyr down or we kind of make a caricature of the city instead of making it more realistic. Im Going to bed now.